Refactor wallet.secrets to a separate file from crypto.go
